pyproject.toml
[project]
name = "python-utils"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pypdf>=6.0.0",
    "requests>=2.32.5",
    "scipy>=1.16.1",
    "wordninja>=2.0.0",
]
